swarm/storage/mru: refactor and cleanup
