+++ swarm/storage/mru/request.go
@@ -20,19 +20,12 @@ import (
 	"bytes"
 	"encoding/json"
 	"fmt"
-	"hash"
 	"time"
 
 	"github.com/ethereum/go-ethereum/common"
 	"github.com/ethereum/go-ethereum/common/hexutil"
-	"github.com/ethereum/go-ethereum/crypto"
 	"github.com/ethereum/go-ethereum/swarm/storage"
 )
-
-// Signature is an alias for a static byte array with the size of a signature
-const signatureLength = 65
-
-type Signature [signatureLength]byte
 
 // updateRequestJSON represents a JSON-serialized UpdateRequest
 type updateRequestJSON struct {
@@ -49,39 +42,14 @@ type updateRequestJSON struct {
 	Signature string `json:"signature,omitempty"`
 }
 
-// SignedResourceUpdate contains signature information about a resource update
-type SignedResourceUpdate struct {
-	resourceUpdate // actual content that will be put on the chunk, less signature
-	signature      *Signature
-	updateAddr     storage.Address // resulting chunk address for the update
-}
-
-// Update chunk layout
-// Header:
-// 2 bytes header Length
-// 2 bytes data length
-// 4 bytes period
-// 4 bytes version
-// 32 bytes rootAddr reference
-// 32 bytes metaHash digest
-// Data:
-// data (datalength bytes)
-// Signature:
-// signature: 65 bytes (signatureLength constant)
-//
-// Minimum size is Header + 1 (minimum data length, enforced) + Signature = 142 bytes
-
-const minimumUpdateDataLength = 142
-const maxUpdateDataLength = chunkSize - minimumUpdateDataLength
-
-// UpdateRequest represents an update and/or resource create message
-type UpdateRequest struct {
+// Request represents an update and/or resource create message
+type Request struct {
 	SignedResourceUpdate
 	resourceMetadata
 }
 
-// NewCreateRequest returns a ready to sign UpdateRequest message to create a new resource
-func NewCreateRequest(name string, frequency, startTime uint64, ownerAddr common.Address, data []byte, multihash bool) (*UpdateRequest, error) {
+// NewCreateRequest returns a ready to sign Request message to create a new resource
+func NewCreateRequest(name string, frequency, startTime uint64, ownerAddr common.Address, data []byte, multihash bool) (*Request, error) {
 	if !isSafeName(name) {
 		return nil, NewError(ErrInvalidValue, fmt.Sprintf("Invalid name: '%s' when creating a new UpdateRequest", name))
 	}
@@ -91,7 +59,7 @@ func NewCreateRequest(name string, frequency, startTime uint64, ownerAddr common
 		startTime = uint64(time.Now().Unix())
 	}
 
-	updateRequest := &UpdateRequest{
+	request := &Request{
 		SignedResourceUpdate: SignedResourceUpdate{
 			resourceUpdate: resourceUpdate{
 				updateHeader: updateHeader{
@@ -112,53 +80,53 @@ func NewCreateRequest(name string, frequency, startTime uint64, ownerAddr common
 		},
 	}
 
-	updateRequest.rootAddr, updateRequest.metaHash, _ = updateRequest.resourceMetadata.hash()
+	request.rootAddr, request.metaHash, _ = request.resourceMetadata.hash()
 
-	return updateRequest, nil
+	return request, nil
 }
 
 // Frequency Returns the resource expected update frequency
-func (r *UpdateRequest) Frequency() uint64 {
+func (r *Request) Frequency() uint64 {
 	return r.frequency
 }
 
 // Name returns the resource human-readable name
-func (r *UpdateRequest) Name() string {
+func (r *Request) Name() string {
 	return r.name
 }
 
 // Multihash returns true if the resource data should be interpreted as a multihash
-func (r *UpdateRequest) Multihash() bool {
+func (r *Request) Multihash() bool {
 	return r.multihash
 }
 
 // Period returns in which period the resource will be published
-func (r *UpdateRequest) Period() uint32 {
+func (r *Request) Period() uint32 {
 	return r.period
 }
 
 // Version returns the resource version to publish
-func (r *UpdateRequest) Version() uint32 {
+func (r *Request) Version() uint32 {
 	return r.version
 }
 
 // RootAddr returns the metadata chunk address
-func (r *UpdateRequest) RootAddr() storage.Address {
+func (r *Request) RootAddr() storage.Address {
 	return r.rootAddr
 }
 
 // StartTime returns the time that the resource was/will be created at
-func (r *UpdateRequest) StartTime() uint64 {
+func (r *Request) StartTime() uint64 {
 	return r.startTime
 }
 
 // OwnerAddr returns the resource owner's address
-func (r *UpdateRequest) OwnerAddr() common.Address {
+func (r *Request) OwnerAddr() common.Address {
 	return r.ownerAddr
 }
 
 // Sign executes the signature to validate the resource and sets the owner address field
-func (r *UpdateRequest) Sign(signer Signer) error {
+func (r *Request) Sign(signer Signer) error {
 	if err := r.SignedResourceUpdate.Sign(signer); err != nil {
 		return err
 	}
@@ -166,68 +134,17 @@ func (r *UpdateRequest) Sign(signer Signer) error {
 	return nil
 }
 
-// Verify checks that signatures are valid and that the signer owns the resource to be updated
-func (r *SignedResourceUpdate) Verify() (err error) {
-	if len(r.data) == 0 {
-		return NewError(ErrInvalidValue, "I refuse to waste swarm space for updates with empty values, amigo (data length is 0)")
-	}
-	if r.signature == nil {
-		return NewError(ErrInvalidSignature, "Missing signature field")
-	}
-
-	digest := resourceUpdateChunkDigest(r.updateAddr, r.metaHash, r.data)
-
-	// get the address of the signer (which also checks that it's a valid signature)
-	ownerAddr, err := getAddressFromDataSig(digest, *r.signature)
-	if err != nil {
-		return err
-	}
-
-	if !bytes.Equal(r.updateAddr, r.GetUpdateAddr()) {
-		return NewError(ErrInvalidSignature, "Signature address does not match with ownerAddr")
-	}
-
-	// Check if who signed the resource update really owns the resource
-	if !verifyResourceOwnership(ownerAddr, r.metaHash, r.rootAddr) {
-		return NewError(ErrUnauthorized, fmt.Sprintf("signature is valid but signer does not own the resource: %v", err))
-	}
-
-	return nil
-}
-
-// Sign executes the signature to validate the resource
-func (r *SignedResourceUpdate) Sign(signer Signer) error {
-
-	updateAddr := r.GetUpdateAddr()
-
-	digest := resourceUpdateChunkDigest(updateAddr, r.metaHash, r.data)
-	signature, err := signer.Sign(digest)
-	if err != nil {
-		return err
-	}
-	ownerAddress, err := getAddressFromDataSig(digest, signature)
-	if err != nil {
-		return NewError(ErrInvalidSignature, "Error verifying signature")
-	}
-	if ownerAddress != signer.Address() {
-		return NewError(ErrInvalidSignature, "Signer address does not match private key")
-	}
-	r.signature = &signature
-	r.updateAddr = updateAddr
-	return nil
-}
-
 // SetData stores the payload data the resource will be updated with
-func (r *UpdateRequest) SetData(data []byte) {
+func (r *Request) SetData(data []byte) {
 	r.signature = nil
 	r.data = data
 	r.frequency = 0 //mark as update
 }
 
 // decode takes an update request JSON and returns an UpdateRequest
-func (j *updateRequestJSON) decode() (*UpdateRequest, error) {
+func (j *updateRequestJSON) decode() (*Request, error) {
 
-	r := &UpdateRequest{
+	r := &Request{
 		SignedResourceUpdate: SignedResourceUpdate{
 			resourceUpdate: resourceUpdate{
 				updateHeader: updateHeader{
@@ -314,7 +231,7 @@ func (j *updateRequestJSON) decode() (*UpdateRequest, error) {
 }
 
 // DecodeUpdateRequest takes a JSON structure stored in a byte array and returns a live UpdateRequest object
-func DecodeUpdateRequest(rawData []byte) (*UpdateRequest, error) {
+func DecodeUpdateRequest(rawData []byte) (*Request, error) {
 	var requestJSON updateRequestJSON
 	if err := json.Unmarshal(rawData, &requestJSON); err != nil {
 		return nil, err
@@ -323,7 +240,7 @@ func DecodeUpdateRequest(rawData []byte) (*UpdateRequest, error) {
 }
 
 // EncodeUpdateRequest takes an update request and encodes it as a JSON structure into a byte array
-func EncodeUpdateRequest(updateRequest *UpdateRequest) (rawData []byte, err error) {
+func EncodeUpdateRequest(updateRequest *Request) (rawData []byte, err error) {
 	var signatureString, dataHashString, rootAddrString, metaHashString string
 	if updateRequest.signature != nil {
 		signatureString = hexutil.Encode(updateRequest.signature[:])
@@ -359,34 +276,4 @@ func EncodeUpdateRequest(updateRequest *UpdateRequest) (rawData []byte, err erro
 	}
 
 	return json.Marshal(requestJSON)
-}
-
-// resourceUpdateChunkDigest creates the resource update digest used in signatures (formerly known as keyDataHash)
-func resourceUpdateChunkDigest(updateAddr storage.Address, metaHash []byte, data []byte) common.Hash {
-	hasher := hashPool.Get().(hash.Hash)
-	defer hashPool.Put(hasher)
-	hasher.Reset()
-	hasher.Write(updateAddr)
-	hasher.Write(metaHash)
-	hasher.Write(data)
-	return common.BytesToHash(hasher.Sum(nil))
-}
-
-// getAddressFromDataSig extracts the address of the resource update signer
-func getAddressFromDataSig(digest common.Hash, signature Signature) (common.Address, error) {
-	pub, err := crypto.SigToPub(digest.Bytes(), signature[:])
-	if err != nil {
-		return common.Address{}, err
-	}
-	return crypto.PubkeyToAddress(*pub), nil
-}
-
-func verifyResourceOwnership(ownerAddr common.Address, metaHash []byte, rootAddr storage.Address) bool {
-	hasher := hashPool.Get().(hash.Hash)
-	defer hashPool.Put(hasher)
-	hasher.Reset()
-	hasher.Write(metaHash)
-	hasher.Write(ownerAddr.Bytes())
-	rootAddr2 := hasher.Sum(nil)
-	return bytes.Equal(rootAddr2, rootAddr)
 }
