+++ swarm/storage/mru/resource.go
@@ -20,12 +20,9 @@ import (
 	"bytes"
 	"context"
 	"encoding/binary"
-	"errors"
 	"fmt"
 	"math/big"
-	"sync"
 	"time"
-	"unsafe"
 
 	"golang.org/x/net/idna"
 
@@ -33,7 +30,6 @@ import (
 
 	"github.com/ethereum/go-ethereum/core/types"
 	"github.com/ethereum/go-ethereum/swarm/log"
-	"github.com/ethereum/go-ethereum/swarm/multihash"
 	"github.com/ethereum/go-ethereum/swarm/storage"
 )
 
@@ -73,38 +69,6 @@ func (b *blockEstimator) HeaderByNumber(context.Context, string, *big.Int) (*typ
 	return &types.Header{
 		Number: big.NewInt(time.Since(b.Start).Nanoseconds() / b.Average.Nanoseconds()),
 	}, nil
-}
-
-// Error is a the typed error object used for Mutable Resources
-type Error struct {
-	code int
-	err  string
-}
-
-// Error implements the error interface
-func (e *Error) Error() string {
-	return e.err
-}
-
-// Code returns the error code
-// Error codes are enumerated in the error.go file within the mru package
-func (e *Error) Code() int {
-	return e.code
-}
-
-// NewError creates a new Mutable Resource Error object with the specified code and custom error message
-func NewError(code int, s string) error {
-	if code < 0 || code >= ErrCnt {
-		panic("no such error code!")
-	}
-	r := &Error{
-		err: s,
-	}
-	switch code {
-	case ErrNotFound, ErrIO, ErrUnauthorized, ErrInvalidValue, ErrDataOverflow, ErrNothingToReturn, ErrInvalidSignature, ErrNotSynced, ErrPeriodDepth, ErrCorruptData:
-		r.code = code
-	}
-	return r
 }
 
 type updateHeader struct {
@@ -154,687 +118,6 @@ func (r *resource) Size(ctx context.Context, _ chan bool) (int64, error) {
 //returns the resource's human-readable name
 func (r *resource) Name() string {
 	return r.name
-}
-
-// Handler is the API for Mutable Resources
-// It enables creating, updating, syncing and retrieving resources and their update data
-type Handler struct {
-	chunkStore        *storage.NetStore
-	HashSize          int
-	timestampProvider timestampProvider
-	resources         map[uint64]*resource
-	resourceLock      sync.RWMutex
-	storeTimeout      time.Duration
-	queryMaxPeriods   uint32
-}
-
-// HandlerParams pass parameters to the Handler constructor NewHandler
-// Signer and TimestampProvider are mandatory parameters
-type HandlerParams struct {
-	QueryMaxPeriods   uint32
-	TimestampProvider timestampProvider
-}
-
-// hashPool contains a pool of ready hashers
-var hashPool sync.Pool
-
-// init initializes the package and hashPool
-func init() {
-	hashPool = sync.Pool{
-		New: func() interface{} {
-			return storage.MakeHashFunc(resourceHashAlgorithm)()
-		},
-	}
-}
-
-// NewHandler creates a new Mutable Resource API
-func NewHandler(params *HandlerParams) (*Handler, error) {
-
-	rh := &Handler{
-		timestampProvider: params.TimestampProvider,
-		resources:         make(map[uint64]*resource),
-		storeTimeout:      defaultStoreTimeout,
-		queryMaxPeriods:   params.QueryMaxPeriods,
-	}
-
-	if rh.timestampProvider == nil {
-		rh.timestampProvider = NewDefaultTimestampProvider()
-	}
-
-	for i := 0; i < hasherCount; i++ {
-		hashfunc := storage.MakeHashFunc(resourceHashAlgorithm)()
-		if rh.HashSize == 0 {
-			rh.HashSize = hashfunc.Size()
-		}
-		hashPool.Put(hashfunc)
-	}
-
-	return rh, nil
-}
-
-// SetStore sets the store backend for the Mutable Resource API
-func (h *Handler) SetStore(store *storage.NetStore) {
-	h.chunkStore = store
-}
-
-// Validate is a chunk validation method
-// If it's a resource update, the chunk address is checked against the ownerAddr of the update's signature
-// It implements the storage.ChunkValidator interface
-func (h *Handler) Validate(chunkAddr storage.Address, data []byte) bool {
-
-	//metadata chunks have the first two bytes set to zero
-	if data[0] == 0 && data[1] == 0 && len(data) > common.AddressLength {
-		//metadata chunk
-		rootAddr, _ := metadataHash(data)
-		valid := bytes.Equal(chunkAddr, rootAddr)
-		if !valid {
-			log.Warn("Invalid root metadata chunk")
-		}
-		return valid
-	}
-
-	// if it is not a metadata chunk, check if it is an update chunk with
-	// valid signature and proof of ownership of the resource it is trying
-	// to update
-
-	_, err := parseUpdate(chunkAddr, data)
-	if err != nil {
-		log.Warn("Invalid resource chunk")
-		return false
-	}
-
-	return true
-}
-
-// GetContent retrieves the data payload of the last synced update of the Mutable Resource
-func (h *Handler) GetContent(rootAddr storage.Address) (storage.Address, []byte, error) {
-	rsrc := h.get(rootAddr)
-	if rsrc == nil || !rsrc.isSynced() {
-		return nil, nil, NewError(ErrNotFound, " does not exist or is not synced")
-	}
-	return rsrc.lastKey, rsrc.data, nil
-}
-
-// GetLastPeriod retrieves the period of the last synced update of the Mutable Resource
-func (h *Handler) GetLastPeriod(rootAddr storage.Address) (uint32, error) {
-	rsrc := h.get(rootAddr)
-	if rsrc == nil {
-		return 0, NewError(ErrNotFound, " does not exist")
-	} else if !rsrc.isSynced() {
-		return 0, NewError(ErrNotSynced, " is not synced")
-	}
-	return rsrc.period, nil
-}
-
-// GetVersion retrieves the period of the last synced update of the Mutable Resource
-func (h *Handler) GetVersion(rootAddr storage.Address) (uint32, error) {
-	rsrc := h.get(rootAddr)
-	if rsrc == nil {
-		return 0, NewError(ErrNotFound, " does not exist")
-	} else if !rsrc.isSynced() {
-		return 0, NewError(ErrNotSynced, " is not synced")
-	}
-	return rsrc.version, nil
-}
-
-// \TODO should be hashsize * branches from the chosen chunker, implement with FileStore
-func (h *Handler) chunkSize() int64 {
-	return chunkSize
-}
-
-// New creates a new metadata chunk out of the request passed in.
-func (h *Handler) New(ctx context.Context, request *UpdateRequest) error {
-
-	// frequency 0 is invalid
-	if request.frequency == 0 {
-		return NewError(ErrInvalidValue, "frequency cannot be 0 when creating a resource")
-	}
-
-	// make sure name only contains ascii values
-	if !isSafeName(request.name) {
-		return NewError(ErrInvalidValue, fmt.Sprintf("invalid name: '%s'", request.name))
-	}
-
-	// make sure owner is set to something
-	var zeroAddr = common.Address{}
-	if request.ownerAddr == zeroAddr {
-		return NewError(ErrInvalidValue, "ownerAddr must be set to create a new metadata chunk")
-	}
-
-	// create the meta chunk and store it in swarm
-	chunk, metaHash := h.newMetaChunk(&request.resourceMetadata)
-	if request.metaHash != nil && !bytes.Equal(request.metaHash, metaHash) ||
-		request.rootAddr != nil && !bytes.Equal(request.rootAddr, chunk.Addr) {
-		return NewError(ErrInvalidValue, "metaHash in UpdateRequest does not match actual metadata")
-	}
-
-	request.metaHash = metaHash
-	request.rootAddr = chunk.Addr
-
-	h.chunkStore.Put(ctx, chunk)
-	log.Debug("new resource", "name", request.name, "startBlock", request.startTime, "frequency", request.frequency, "owner", request.ownerAddr)
-
-	// create the internal index for the resource and populate it with the data of the first version
-	rsrc := &resource{
-		resourceUpdate: resourceUpdate{
-			updateHeader: updateHeader{
-				UpdateLookup: UpdateLookup{
-					rootAddr: chunk.Addr,
-				},
-			},
-		},
-		resourceMetadata: resourceMetadata{
-			name:      request.name,
-			startTime: request.startTime,
-			frequency: request.frequency,
-		},
-		updated: time.Now(),
-	}
-	copy(rsrc.ownerAddr[:], request.ownerAddr[:])
-	h.set(chunk.Addr, rsrc)
-
-	return nil
-}
-
-// NewUpdateRequest prepares an UpdateRequest structure with all the necessary information to
-// just add the desired data and sign it.
-// The resulting structure can then be signed and passed to Handler.Update to be verified and sent
-func (h *Handler) NewUpdateRequest(ctx context.Context, rootAddr storage.Address) (*UpdateRequest, error) {
-
-	if rootAddr == nil {
-		return nil, NewError(ErrInvalidValue, "rootAddr cannot be nil")
-	}
-
-	rsrc, err := h.Load(ctx, rootAddr)
-	if err != nil {
-		return nil, err
-	}
-
-	currentblock := h.getCurrentTime(ctx)
-
-	updateRequest := new(UpdateRequest)
-	updateRequest.period, err = getNextPeriod(rsrc.startTime, currentblock, rsrc.frequency)
-	if err != nil {
-		return nil, err
-	}
-	if _, err = h.lookup(rsrc, LookupLatestVersionInPeriod(rsrc.rootAddr, updateRequest.period)); err != nil {
-		return nil, err
-	}
-
-	if !rsrc.isSynced() {
-		return nil, NewError(ErrNotSynced, fmt.Sprintf("Handler.NewUpdateRequest: object '%s' not in sync", rootAddr.Hex()))
-	}
-
-	updateRequest.multihash = rsrc.multihash
-	updateRequest.rootAddr = rsrc.rootAddr
-	updateRequest.metaHash = rsrc.metaHash
-	updateRequest.resourceMetadata = rsrc.resourceMetadata
-
-	// if we already have an update for this block then increment version
-	// resource object MUST be in sync for version to be correct, but we checked this earlier in the method already
-	if h.hasUpdate(rootAddr, updateRequest.period) {
-		updateRequest.version = rsrc.version + 1
-	} else {
-		updateRequest.version = 1
-	}
-
-	return updateRequest, nil
-}
-
-// creates a metadata chunk out of a resourceMetadata structure
-//TODO: refactor to a method of the resourceMetadata structure
-func (h *Handler) newMetaChunk(metadata *resourceMetadata) (chunk *storage.Chunk, metaHash []byte) {
-	// the metadata chunk contains a timestamp of when the resource starts to be valid
-	// and also how frequently it is expected to be updated
-	// from this we know at what time we should look for updates, and how often
-	// it also contains the name of the resource, so we know what resource we are working with
-
-	// the key (rootAddr) of the metadata chunk is content-addressed
-	// if it wasn't we couldn't replace it later
-	// resolving this relationship is left up to external agents (for example ENS)
-	rootAddr, metaHash, chunkData := metadata.hash()
-
-	// make the chunk and send it to swarm
-	chunk = storage.NewChunk(rootAddr, nil)
-	chunk.SData = chunkData
-	chunk.Size = int64(len(chunkData))
-
-	return chunk, metaHash
-}
-
-// LookupLatest retrieves the latest version of the resource update with metadata chunk at params.Root
-// It starts at the next period after the current block height, and upon failure
-// tries the corresponding keys of each previous period until one is found
-// (or startBlock is reached, in which case there are no updates).
-func (h *Handler) Lookup(ctx context.Context, params *LookupParams) (*resource, error) {
-
-	rsrc := h.get(params.rootAddr)
-	if rsrc == nil {
-		return nil, NewError(ErrNothingToReturn, "resource not loaded")
-	}
-	if params.period == 0 {
-		// get our blockheight at this time and the next block of the update period
-		currentblock := h.getCurrentTime(ctx)
-
-		var period uint32
-		period, err := getNextPeriod(rsrc.startTime, currentblock, rsrc.frequency)
-		if err != nil {
-			return nil, err
-		}
-		params.period = period
-	}
-	return h.lookup(rsrc, params)
-}
-
-// LookupPreviousByName returns the resource before the one currently loaded in the resource index
-// This is useful where resource updates are used incrementally in contrast to
-// merely replacing content.
-// Requires a synced resource object
-func (h *Handler) LookupPrevious(ctx context.Context, params *LookupParams) (*resource, error) {
-	rsrc := h.get(params.rootAddr)
-	if rsrc == nil {
-		return nil, NewError(ErrNothingToReturn, "resource not loaded")
-	}
-	if !rsrc.isSynced() {
-		return nil, NewError(ErrNotSynced, "LookupPrevious requires synced resource.")
-	} else if rsrc.period == 0 {
-		return nil, NewError(ErrNothingToReturn, " not found")
-	}
-	if rsrc.version > 1 {
-		rsrc.version--
-	} else if rsrc.period == 1 {
-		return nil, NewError(ErrNothingToReturn, "Current update is the oldest")
-	} else {
-		rsrc.version = 0
-		rsrc.period--
-	}
-	return h.lookup(rsrc, NewLookupParams(rsrc.rootAddr, rsrc.period, rsrc.version, params.Limit))
-}
-
-// base code for public lookup methods
-func (h *Handler) lookup(rsrc *resource, params *LookupParams) (*resource, error) {
-
-	lp := *params
-	// we can't look for anything without a store
-	if h.chunkStore == nil {
-		return nil, NewError(ErrInit, "Call Handler.SetStore() before performing lookups")
-	}
-
-	// period 0 does not exist
-	if lp.period == 0 {
-		return nil, NewError(ErrInvalidValue, "period must be >0")
-	}
-
-	// start from the last possible block period, and iterate previous ones until we find a match
-	// if we hit startBlock we're out of options
-	var specificversion bool
-	if lp.version > 0 {
-		specificversion = true
-	} else {
-		lp.version = 1
-	}
-
-	var hops uint32
-	if lp.Limit == 0 {
-		lp.Limit = h.queryMaxPeriods
-	}
-	log.Trace("resource lookup", "period", lp.period, "version", lp.version, "limit", lp.Limit)
-	for lp.period > 0 {
-		if lp.Limit != 0 && hops > lp.Limit {
-			return nil, NewError(ErrPeriodDepth, fmt.Sprintf("Lookup exceeded max period hops (%d)", lp.Limit))
-		}
-		updateAddr := lp.GetUpdateAddr()
-		chunk, err := h.chunkStore.GetWithTimeout(context.TODO(), updateAddr, defaultRetrieveTimeout)
-		if err == nil {
-			if specificversion {
-				return h.updateIndex(rsrc, chunk)
-			}
-			// check if we have versions > 1. If a version fails, the previous version is used and returned.
-			log.Trace("rsrc update version 1 found, checking for version updates", "period", lp.period, "updateAddr", updateAddr)
-			for {
-				newversion := lp.version + 1
-				updateAddr := lp.GetUpdateAddr()
-				newchunk, err := h.chunkStore.GetWithTimeout(context.TODO(), updateAddr, defaultRetrieveTimeout)
-				if err != nil {
-					return h.updateIndex(rsrc, chunk)
-				}
-				chunk = newchunk
-				lp.version = newversion
-				log.Trace("version update found, checking next", "version", lp.version, "period", lp.period, "updateAddr", updateAddr)
-			}
-		}
-		log.Trace("rsrc update not found, checking previous period", "period", lp.period, "updateAddr", updateAddr)
-		lp.period--
-		hops++
-	}
-	return nil, NewError(ErrNotFound, "no updates found")
-}
-
-// Load retrieves the Mutable Resource metadata chunk stored at rootAddr
-// Upon retrieval it creates/updates the index entry for it with metadata corresponding to the chunk contents
-func (h *Handler) Load(ctx context.Context, rootAddr storage.Address) (*resource, error) {
-	chunk, err := h.chunkStore.GetWithTimeout(ctx, rootAddr, defaultRetrieveTimeout)
-	if err != nil {
-		return nil, NewError(ErrNotFound, err.Error())
-	}
-
-	// \TODO this is not enough to make sure the data isn't bogus. A normal content addressed chunk could still satisfy these criteria
-	if len(chunk.SData) <= metadataChunkOffsetSize {
-		return nil, NewError(ErrNothingToReturn, fmt.Sprintf("Invalid chunk length %d, should be minimum %d", len(chunk.SData), metadataChunkOffsetSize+1))
-	}
-
-	// create the index entry
-	rsrc := &resource{}
-	rsrc.unmarshalBinary(chunk.SData)
-	rsrc.rootAddr, rsrc.metaHash = metadataHash(chunk.SData)
-	if !bytes.Equal(rsrc.rootAddr, rootAddr) {
-		return nil, NewError(ErrCorruptData, "Corrupt metadata chunk")
-	}
-	h.set(rootAddr, rsrc)
-	log.Trace("resource index load", "rootkey", rootAddr, "name", rsrc.name, "startblock", rsrc.startTime, "frequency", rsrc.frequency)
-	return rsrc, nil
-}
-
-// update mutable resource index map with specified content
-func (h *Handler) updateIndex(rsrc *resource, chunk *storage.Chunk) (*resource, error) {
-
-	// retrieve metadata from chunk data and check that it matches this mutable resource
-	mru, err := parseUpdate(chunk.Addr, chunk.SData)
-	if err != nil {
-		return nil, NewError(ErrInvalidSignature, fmt.Sprintf("Invalid resource chunk: %s", err))
-	}
-	log.Trace("resource index update", "name", rsrc.name, "updatekey", chunk.Addr, "period", mru.period, "version", mru.version)
-
-	// update our rsrcs entry map
-	rsrc.lastKey = chunk.Addr
-	rsrc.period = mru.period
-	rsrc.version = mru.version
-	rsrc.updated = time.Now()
-	rsrc.data = make([]byte, len(mru.data))
-	rsrc.multihash = mru.multihash
-	rsrc.Reader = bytes.NewReader(rsrc.data)
-	copy(rsrc.data, mru.data)
-	log.Debug(" synced", "name", rsrc.name, "updateAddr", chunk.Addr, "period", rsrc.period, "version", rsrc.version)
-	h.set(chunk.Addr, rsrc)
-	return rsrc, nil
-}
-
-// retrieve update metadata from chunk data
-// mirrors newUpdateChunk(). TODO: convert to a SignedResourceUpdate method.
-func parseUpdate(updateAddr storage.Address, chunkdata []byte) (*SignedResourceUpdate, error) {
-	// for update chunk layout see SignedResourceUpdate definition
-
-	if len(chunkdata) < minimumUpdateDataLength {
-		return nil, NewError(ErrNothingToReturn, fmt.Sprintf("chunk less than %d bytes cannot be a resource update chunk", minimumUpdateDataLength))
-	}
-	cursor := 0
-	headerlength := binary.LittleEndian.Uint16(chunkdata[cursor : cursor+2])
-	cursor += 2
-	datalength := binary.LittleEndian.Uint16(chunkdata[cursor : cursor+2])
-	cursor += 2
-
-	if datalength != 0 && int(2+2+headerlength+datalength+signatureLength) != len(chunkdata) {
-		return nil, NewError(ErrNothingToReturn, "length specified in header is different than actual chunk size")
-	}
-
-	var exclsignlength int
-	// we need extra magic if it's a multihash, since we used datalength 0 in header as an indicator of multihash content
-	// retrieve the second varint and set this as the data length
-	// TODO: merge with isMultihash code
-	if datalength == 0 {
-		uvarintbuf := bytes.NewBuffer(chunkdata[headerlength+4:])
-		r, err := binary.ReadUvarint(uvarintbuf)
-		if err != nil {
-			errstr := fmt.Sprintf("corrupt multihash, hash id varint could not be read: %v", err)
-			log.Warn(errstr)
-			return nil, NewError(ErrCorruptData, errstr)
-
-		}
-		r, err = binary.ReadUvarint(uvarintbuf)
-		if err != nil {
-			errstr := fmt.Sprintf("corrupt multihash, hash length field could not be read: %v", err)
-			log.Warn(errstr)
-			return nil, NewError(ErrCorruptData, errstr)
-
-		}
-		exclsignlength = int(headerlength + uint16(r))
-	} else {
-		exclsignlength = int(headerlength + datalength + 4)
-	}
-
-	// the total length excluding signature is headerlength and datalength fields plus the length of the header and the data given in these fields
-	exclsignlength = int(headerlength + datalength + 4)
-	if exclsignlength > len(chunkdata) || exclsignlength < 14 {
-		return nil, NewError(ErrNothingToReturn, fmt.Sprintf("Reported headerlength %d + datalength %d longer than actual chunk data length %d", headerlength, exclsignlength, len(chunkdata)))
-	} else if exclsignlength < 14 {
-		return nil, NewError(ErrNothingToReturn, fmt.Sprintf("Reported headerlength %d + datalength %d is smaller than minimum valid resource chunk length %d", headerlength, datalength, 14))
-	}
-
-	// at this point we can be satisfied that the data integrity is ok
-	var header updateHeader
-
-	var data []byte
-	header.period = binary.LittleEndian.Uint32(chunkdata[cursor : cursor+4])
-	cursor += 4
-	header.version = binary.LittleEndian.Uint32(chunkdata[cursor : cursor+4])
-	cursor += 4
-
-	header.rootAddr = storage.Address(make([]byte, storage.KeyLength))
-	header.metaHash = make([]byte, storage.KeyLength)
-	copy(header.rootAddr, chunkdata[cursor:cursor+storage.KeyLength])
-	cursor += storage.KeyLength
-	copy(header.metaHash, chunkdata[cursor:cursor+storage.KeyLength])
-	cursor += storage.KeyLength
-
-	//Verify that the updateAddr key that identifies this chunk matches its contents:
-	if !bytes.Equal(updateAddr, header.GetUpdateAddr()) {
-		return nil, NewError(ErrInvalidSignature, "period,version,rootAddr contained in update chunk do not match updateAddr")
-	}
-
-	// if multihash content is indicated we check the validity of the multihash
-	// \TODO the check above for multihash probably is sufficient also for this case (or can be with a small adjustment) and if so this code should be removed
-	var intdatalength int
-	var ismultihash bool
-	if datalength == 0 {
-		var intheaderlength int
-		var err error
-		intdatalength, intheaderlength, err = multihash.GetMultihashLength(chunkdata[cursor:])
-		if err != nil {
-			log.Error("multihash parse error", "err", err)
-			return nil, err
-		}
-		intdatalength += intheaderlength
-		multihashboundary := cursor + intdatalength
-		if len(chunkdata) != multihashboundary && len(chunkdata) < multihashboundary+signatureLength {
-			log.Debug("multihash error", "chunkdatalen", len(chunkdata), "multihashboundary", multihashboundary)
-			return nil, errors.New("Corrupt multihash data")
-		}
-		ismultihash = true
-	} else {
-		intdatalength = int(datalength)
-	}
-	data = make([]byte, intdatalength)
-	copy(data, chunkdata[cursor:cursor+intdatalength])
-
-	// omit signatures if we have no validator
-	var signature *Signature
-	cursor += intdatalength
-	sigdata := chunkdata[cursor : cursor+signatureLength]
-	if len(sigdata) > 0 {
-		signature = &Signature{}
-		copy(signature[:], sigdata)
-	}
-
-	header.multihash = ismultihash
-
-	r := &SignedResourceUpdate{
-		signature:  signature,
-		updateAddr: updateAddr,
-		resourceUpdate: resourceUpdate{
-			updateHeader: header,
-			data:         data,
-		},
-	}
-
-	if err := r.Verify(); err != nil {
-		return nil, NewError(ErrUnauthorized, fmt.Sprintf("Invalid signature: %v", err))
-	}
-
-	return r, nil
-
-}
-
-// Update adds an actual data update
-// Uses the Mutable Resource metadata currently loaded in the resources map entry.
-// It is the caller's responsibility to make sure that this data is not stale.
-// Note that a Mutable Resource update cannot span chunks, and thus has a MAX NET LENGTH 4096, INCLUDING update header data and signature. An error will be returned if the total length of the chunk payload will exceed this limit.
-func (h *Handler) Update(ctx context.Context, rootAddr storage.Address, mru *SignedResourceUpdate) (storage.Address, error) {
-	return h.update(ctx, rootAddr, mru)
-}
-
-// create and commit an update
-func (h *Handler) update(ctx context.Context, rootAddr storage.Address, mru *SignedResourceUpdate) (storage.Address, error) {
-
-	if mru.multihash && isMultihash(mru.data) == 0 {
-		return nil, NewError(ErrNothingToReturn, "Invalid multihash")
-	}
-
-	// we can't update anything without a store
-	if h.chunkStore == nil {
-		return nil, NewError(ErrInit, "Call Handler.SetStore() before updating")
-	}
-
-	// get the cached information
-
-	rsrc := h.get(rootAddr)
-	if rsrc == nil {
-		return nil, NewError(ErrNotFound, fmt.Sprintf(" object '%s' not in index", rsrc.name))
-	} else if !rsrc.isSynced() {
-		return nil, NewError(ErrNotSynced, " object not in sync")
-	}
-
-	// an update can be only one chunk long; data length less header and signature data
-	if int64(len(mru.data)) > maxUpdateDataLength {
-		return nil, NewError(ErrDataOverflow, fmt.Sprintf("Data overflow: %d / %d bytes", len(mru.data), maxUpdateDataLength))
-	}
-
-	if rsrc.period == mru.period {
-		if mru.version != rsrc.version+1 {
-			return nil, NewError(ErrInvalidValue, fmt.Sprintf("Invalid version for this period. Expected version=%d", rsrc.version+1))
-		}
-	} else {
-		if !(mru.period > rsrc.period && mru.version == 1) {
-			return nil, NewError(ErrInvalidValue, fmt.Sprintf("Invalid version,period. Expected version=1 and period > %d", rsrc.period))
-		}
-	}
-
-	chunk, err := newUpdateChunk(mru)
-	if err != nil {
-		return nil, err
-	}
-
-	// send the chunk
-	h.chunkStore.Put(ctx, chunk)
-	log.Trace("resource update", "updateAddr", mru.updateAddr, "lastperiod", mru.period, "version", mru.version, "data", chunk.SData, "multihash", mru.multihash)
-
-	// update our resources map entry and return the new updateAddr
-	rsrc.period = mru.period
-	rsrc.version = mru.version
-	rsrc.data = make([]byte, len(mru.data))
-	copy(rsrc.data, mru.data)
-	return mru.updateAddr, nil
-}
-
-// gets the current time
-func (h *Handler) getCurrentTime(ctx context.Context) uint64 {
-	return h.timestampProvider.GetCurrentTime()
-}
-
-// Retrieves the resource index value for the given nameHash
-func (h *Handler) get(rootAddr storage.Address) *resource {
-	if len(rootAddr) < storage.KeyLength {
-		log.Warn("Handler.get with invalid rootAddr")
-		return nil
-	}
-	hashKey := *(*uint64)(unsafe.Pointer(&rootAddr[0]))
-	h.resourceLock.RLock()
-	defer h.resourceLock.RUnlock()
-	rsrc := h.resources[hashKey]
-	return rsrc
-}
-
-// Sets the resource index value for the given nameHash
-func (h *Handler) set(rootAddr storage.Address, rsrc *resource) {
-	if len(rootAddr) < storage.KeyLength {
-		log.Warn("Handler.set with invalid rootAddr")
-		return
-	}
-	hashKey := *(*uint64)(unsafe.Pointer(&rootAddr[0]))
-	h.resourceLock.Lock()
-	defer h.resourceLock.Unlock()
-	h.resources[hashKey] = rsrc
-}
-
-// Checks if we already have an update on this resource, according to the value in the current state of the resource index
-func (h *Handler) hasUpdate(rootAddr storage.Address, period uint32) bool {
-	rsrc := h.get(rootAddr)
-	return rsrc != nil && rsrc.period == period
-}
-
-// create an update chunk. //TODO: Move to a method of SignedUpdateResource
-func newUpdateChunk(mru *SignedResourceUpdate) (*storage.Chunk, error) {
-
-	if len(mru.rootAddr) != storage.KeyLength || len(mru.metaHash) != storage.KeyLength {
-		log.Warn("Call to newUpdateChunk with nil rootAddr or metaHash")
-		return nil, NewError(ErrInvalidValue, "newUpdateChunk called without rootAddr or metaHash set")
-	}
-	// a datalength field set to 0 means the content is a multihash
-	var datalength int
-	if !mru.multihash {
-		datalength = len(mru.data)
-		if datalength == 0 {
-			return nil, NewError(ErrInvalidValue, "cannot update a resource with no data")
-		}
-	}
-
-	// prepend version, period, metaHash and rootAddr references
-	headerlength := 4 + 4 + storage.KeyLength + storage.KeyLength
-
-	actualdatalength := len(mru.data)
-	chunk := storage.NewChunk(mru.updateAddr, nil)
-	chunk.SData = make([]byte, 2+2+signatureLength+headerlength+actualdatalength) // initial 4 are uint16 length descriptors for headerlength and datalength
-
-	// data header length does NOT include the header length prefix bytes themselves
-	cursor := 0
-	binary.LittleEndian.PutUint16(chunk.SData[cursor:], uint16(headerlength))
-	cursor += 2
-
-	// data length
-	binary.LittleEndian.PutUint16(chunk.SData[cursor:], uint16(datalength))
-	cursor += 2
-
-	// header = period + version + rootAddr + metaHash
-	binary.LittleEndian.PutUint32(chunk.SData[cursor:], mru.period)
-	cursor += 4
-
-	binary.LittleEndian.PutUint32(chunk.SData[cursor:], mru.version)
-	cursor += 4
-
-	copy(chunk.SData[cursor:], mru.rootAddr[:storage.KeyLength])
-	cursor += storage.KeyLength
-	copy(chunk.SData[cursor:], mru.metaHash[:storage.KeyLength])
-	cursor += storage.KeyLength
-
-	// add the data
-	copy(chunk.SData[cursor:], mru.data)
-
-	// signature is the last item in the chunk data
-
-	cursor += actualdatalength
-	copy(chunk.SData[cursor:], mru.signature[:])
-
-	chunk.Size = int64(len(chunk.SData))
-	return chunk, nil
 }
 
 // Helper function to calculate the next update period number from the current time, start time and frequency
